fix categrories in dashboard && app > add  routes, add cart funcionality, add sticked nav bar,

+++ app/lib/models/sku.dart
@@ -12,7 +12,7 @@ part 'sku.g.dart';
 @JsonSerializable(explicitToJson: true)
 class Sku {
   final String id;
-  final double qty;
+  final int qty;
   final String? hashedColor;
   final String? nameOfColor;
   final String? productId;
